black & split paths.py into paths folder
